refactor(hell-work-job-searcher): 求人情報の無限スクロールAPIに腐食防止層を追加 - tanstack queryにべっとり依存を避けるため - https://speakerdeck.com/okashoi/interface-and-anti-corruption-layer
